Fix performance issue in get_poetry_venv_executable_path
The above function executed the `poetry env` command every time
the `BufRead` event was getting fired which I noticed when I used
telescope to jump to an LSP definition.

The fix was to just check that the executable exists in the `.venv/bin`
directory within the current workspace and just use it.

I had to revert back through my last few commits to debug this issue.