Remove exclude config from basedpyright
For some reason the basedpyright LSP was struggling with certain
projects and the LSP was under heavy CPU load. I spent some time
debugging the issue and I found that the culprit was the `exclude`
config I have for basedpyright. This might have been a hidden problem
for some time, or it might have been introduced with recent LSP updates.